CD-226 refact: 시큐리티 빈 생성 수정
